+++ src/lyskal/module9/evaluationofstudents/Evaluation.java
@@ -1,4 +1,4 @@
-package lyskal.module9;
+package lyskal.module9.evaluationofstudents;
 
 public enum Evaluation {
 	ONE,
